+++ frameworks/native/audiorenderer/include/audio_renderer_proxy_obj.h
@@ -30,7 +30,7 @@ public:
     void ResumeStreamImpl(const StreamSetStateEventInternal &streamSetStateEventInternal);
     void SetLowPowerVolumeImpl(float volume);
     void GetLowPowerVolumeImpl(float &volume);
-    void GetSingleStreamVolumeImpl(uint32_t &volume);
+    void GetSingleStreamVolumeImpl(float &volume);
 private:
     const AudioRenderer *renderer;
 };
