+++ frameworks/native/audiorenderer/src/audio_renderer.cpp
@@ -717,7 +717,7 @@ float AudioRendererPrivate::GetLowPowerVolume() const
     return audioStream_->GetLowPowerVolume();
 }
 
-uint32_t AudioRendererPrivate::GetSingleStreamVolume() const
+float AudioRendererPrivate::GetSingleStreamVolume() const
 {
     return audioStream_->GetSingleStreamVolume();
 }
