+++ services/include/audio_service/client/audio_service_client.h
@@ -498,7 +498,7 @@ public:
     bool VerifyClientPermission(const std::string &permissionName, uint32_t appTokenId, int32_t appUid);
     int32_t SetStreamLowPowerVolume(float powerVolumeFactor);
     float GetStreamLowPowerVolume();
-    uint32_t GetSingleStreamVol();
+    float GetSingleStreamVol();
 
     // Audio timer callback
     virtual void OnTimeOut();
