+++ services/src/audio_service/client/audio_system_manager.cpp
@@ -343,7 +343,7 @@ float AudioSystemManager::GetLowPowerVolume(int32_t streamId) const
     return AudioPolicyManager::GetInstance().GetLowPowerVolume(streamId);
 }
 
-uint32_t AudioSystemManager::GetSingleStreamVolume(int32_t streamId) const
+float AudioSystemManager::GetSingleStreamVolume(int32_t streamId) const
 {
     return AudioPolicyManager::GetInstance().GetSingleStreamVolume(streamId);
 }
